pyproject.toml
[project]
name = "better-moondream-demo"
version = "0.1.0"
description = "Chat with your images using Moondream 3 — a Gradio interface for visual question answering, captioning, object detection, and pointing."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=6.6.0",
    "loguru>=0.7.3",
    "moondream>=0.2.0",
    "moondream-station>=0.1.9",
    "pillow>=10.4.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[project.scripts]
moondream-chat = "main:cli"
